move cm tests
