+++ package.json
@@ -43,7 +43,7 @@
     "test": "jasmine --config=jasmine.json",
     "test:unit": "npm test -- test/unit/**/*-spec.js",
     "test:specs": "npm test -- test/specs/**/*-spec.js",
-    "test:integration": "npm test -- test/integration/**/*-spec.js",
+    "test:cm": "npm test -- test/specs/commonmark/**/*-spec.js",
     "test:old": "node test",
     "test:lint": "eslint bin/marked .",
     "bench": "node test --bench",
